refactored stack analysis module
